refactor: / 라우트를 app 그룹으로 이전하고 단일 page.tsx 제거
